fetaure(backend): models for pipeline
